requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.0
Django==5.0.7
django-redis==5.4.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
exceptiongroup==1.2.2
flake8==7.1.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
mypy==1.11.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
ptvsd==4.3.2
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.3.1
pytest-cov==5.0.0
pytest-django==4.8.0
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.1
types-PyYAML==6.0.12.20240808
types-requests==2.32.0.20240712
typing_extensions==4.12.2
urllib3==2.2.2
